:tada: Hard fork from original
The https://brutalistpelican.com is a really clean and competent theme for Pelican. However, for a Church website, I needed it to follow the United Church of Christ brand guidelines. This is a hard fork from the original to meet that end.

I also needed to publish a calendar of events, so I grafted that on and added some support for bits and bobs needed during templating for that as a plugin.